Rework package structure
